Fix prompt being included in input history
When navigating to previous input history and *current input is empty*,
the `backward-sexp' call in function `inf-clojure-get-old-input' would
not respect comint prompt boundary and jump to beginning of line. This
caused the prompt text to be stored in `comint-stored-incomplete-input',
which is annoying and causes even more problems if prompt text is set to
read-only.

This change makes sure that the extracted buffer substring for current
input does not include the prompt, even when input is empty or just
white space.

Also, introduce a regexp to use for `comint-prompt-regexp' which matches
both main prompt and sub prompts, so that boundary is respected when
`comint-use-prompt-regexp' is set to t.

Fixes #35