Calculate response boundaries explicitly
Change the way the response is calculated, not by search-forward the buffer
used for redirection but by parsing the raw response. Taking into consideration
the optional regexps, a list is calculated with beginning, end and prompt
position (the logic is now in inf-clojure--string-boundaries, which can be unit
tested).

This implicitly solves a hidden bug in inf-clojure--process-response, which was
not able to identify the prompt in some cases because of the many
search-forward.

Finally, this new information is logged so that bugs can be detected more
easily.