Avoid throwing an error for frequent operations like completion
This patch introduces an argument to inf-clojure-proc so that client code can
decide when to error out return the processs. This is necessary because while
it is good to communicate an error on "eval" operations, it is not good to have
it on "under the hood" and very frequent ones like completion and get arglists.