[Fix #57] Add `inf-clojure-meta`
#### Summary

This commit addresses https://github.com/clojure-emacs/inf-clojure/issues/57 (Add an interactive
command to display a var's metadata) by adding an interactive command `inf-clojure-meta` to show metadata of a form at point.

This command works with clojure, lumo and planck alike.

#### Test Plan
Manually tested the command using clojure, lumo and planck. I put my cursor on a
form and ran it interactively, it displayed the metadata in the repl buffer.
Specifically I tested with `(with-meta [1 2 3] {:data "hello"})` and it printed `{:data "hello"}`