Set artificial to true on comint-send-input
From comint-send-input:

> After the input has been sent, if `comint-process-echoes' is non-nil,
then `comint-send-input' waits to see if the process outputs a string
matching the input, and if so, deletes that part of the output.
If ARTIFICIAL is non-nil, it inhibits such deletion.
Callers sending input not from the user should use ARTIFICIAL = t.

Seems to mess up a buffer somewhere and output isn't
flushed. Extremely important